Added fixes requested pt2
